+++ fibonacci.h
@@ -2,6 +2,6 @@
 #ifndef FIBONACCI_H
 #define FIBONACCI_H
 
-int fibonacci(int N);
+long long int fibonacci(int N);
 
 #endif // FIBONACCI_H
